requirements.txt
llama-index>=0.8.0
llama-cloud-services>=0.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
pydantic>=2.0.0
python-multipart>=0.0.6
fastapi>=0.95.0
uvicorn>=0.22.0
